pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aria-vrs-extractor"
version = "0.1.0"
description = "Meta Project Aria VRS to JPEG/WAV/JSONL extractor"
readme = "README.md"
authors = [{name = "Project Aria Tools"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "projectaria-tools",
    "typer>=0.9",
    "PyYAML>=6.0",
    "fsspec>=2023.9.0",
    "Pillow>=9.5",
    "numpy>=1.24",
]

[project.scripts]
aria-vrs-extract = "aria_vrs_extractor.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["aria_vrs_extractor*"]
